move pages folder to root of src
