+++ src/scss/styles.scss
@@ -13,7 +13,7 @@ $button-bg: #2f2e33;
 $button-body: #fcfcfc;
 
 // Home
-$banner-image: "../assets/home_banner.png";
+$banner-image: "./assets/home_banner.png";
 
 // Animating
 $fade-duration: 0.2s;
